browser/devtools/webide/test/test_telemetry.html

<!DOCTYPE html>

<html>

  <head>
    <meta charset="utf8">
    <title></title>

    <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
    <script type="application/javascript" src="chrome://mochikit/content/chrome-harness.js"></script>
    <script type="application/javascript;version=1.8" src="head.js"></script>
    <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  </head>

  <body>

    <script type="application/javascript;version=1.8">
      const { _DeprecatedUSBRuntime, _WiFiRuntime, _SimulatorRuntime,
              _gRemoteRuntime, _gLocalRuntime, RuntimeTypes }
            = require("devtools/webide/runtimes");

      // Because we need to gather stats for the period of time that a tool has
      // been opened we make use of setTimeout() to create tool active times.
      const TOOL_DELAY = 200;

      function cycleWebIDE() {
        return Task.spawn(function*() {
          let win = yield openWebIDE();
          // Wait a bit, so we're open for a non-zero time
          yield waitForTime(TOOL_DELAY);
          yield closeWebIDE(win);
        });
      }

      function addFakeRuntimes(win) {
        // We use the real runtimes here (and switch out some functionality)
        // so we can ensure that logging happens as it would in real use.

        let usb = new _DeprecatedUSBRuntime("fakeUSB");
        // Use local pipe instead
        usb.connect = function(connection) {
          ok(connection, win.AppManager.connection, "connection is valid");
          connection.host = null; // force connectPipe
          connection.connect();
          return promise.resolve();
        };
        win.AppManager.runtimeList.usb.push(usb);

        let wifi = new _WiFiRuntime("fakeWiFi");
        // Use local pipe instead
        wifi.connect = function(connection) {
          ok(connection, win.AppManager.connection, "connection is valid");
          connection.host = null; // force connectPipe
          connection.connect();
          return promise.resolve();
        };
        win.AppManager.runtimeList.wifi.push(wifi);

        let sim = new _SimulatorRuntime("fakeSimulator");
        // Use local pipe instead
        sim.connect = function(connection) {
          ok(connection, win.AppManager.connection, "connection is valid");
          connection.host = null; // force connectPipe
          connection.connect();
          return promise.resolve();
        };
        Object.defineProperty(sim, "name", {
          get() {
            return this.version;
          }
        });
        win.AppManager.runtimeList.simulator.push(sim);

        let remote = _gRemoteRuntime;
        // Use local pipe instead
        remote.connect = function(connection) {
          ok(connection, win.AppManager.connection, "connection is valid");
          connection.host = null; // force connectPipe
          connection.connect();
          return promise.resolve();
        };
        let local = _gLocalRuntime;

        let other = Object.create(_gLocalRuntime);
        other.type = RuntimeTypes.OTHER;

        win.AppManager.runtimeList.other = [remote, local, other];

        win.AppManager.update("runtime-list");
      }

      function addTestApp(win) {
        return Task.spawn(function*() {
          let packagedAppLocation = getTestFilePath("app");
          yield win.Cmds.importPackagedApp(packagedAppLocation);
          yield waitForUpdate(win, "project-validated");
        });
      }

      function startConnection(win, type, index) {
        let panelNode = win.document.querySelector("#runtime-panel");
        let items = panelNode.querySelectorAll(".runtime-panel-item-" + type);
        if (index === undefined) {
          is(items.length, 1, "Found one runtime button");
        }

        let deferred = promise.defer();
        win.AppManager.connection.once(
            win.Connection.Events.CONNECTED,
            () => deferred.resolve());

        items[index || 0].click();

        return deferred.promise;
      }

      function waitUntilConnected(win) {
        return Task.spawn(function*() {
          ok(win.document.querySelector("window").className, "busy", "UI is busy");
          yield win.UI._busyPromise;
          is(Object.keys(DebuggerServer._connections).length, 1, "Connected");
        });
      }

      function connectToRuntime(win, type, index) {
        return Task.spawn(function*() {
          startConnection(win, type, index);
          yield waitUntilConnected(win);
        });
      }

      function checkResults() {
        // For help generating these tests use generateTelemetryTests("DEVTOOLS_")
        // here.
        checkTelemetry("DEVTOOLS_DEBUGGER_RDP_LOCAL_LISTTABS_MS", null, "hasentries");
        checkTelemetry("DEVTOOLS_WEBIDE_CONNECTION_DEBUG_USED", [6,0,0]);
        checkTelemetry("DEVTOOLS_WEBIDE_CONNECTION_PLAY_USED", [6,0,0]);
        checkTelemetry("DEVTOOLS_WEBIDE_CONNECTION_RESULT", [0,6,0]);
        checkTelemetry("DEVTOOLS_WEBIDE_CONNECTION_TIME_SECONDS", null, "hasentries");
        checkTelemetry("DEVTOOLS_WEBIDE_LOCAL_CONNECTION_RESULT", [0,1,0]);
        checkTelemetry("DEVTOOLS_WEBIDE_OPENED_BOOLEAN", [0,2,0]);
        checkTelemetry("DEVTOOLS_WEBIDE_OPENED_PER_USER_FLAG", [0,1,0]);
        checkTelemetry("DEVTOOLS_WEBIDE_OTHER_CONNECTION_RESULT", [0,1,0]);
        checkTelemetry("DEVTOOLS_WEBIDE_REMOTE_CONNECTION_RESULT", [0,1,0]);
        checkTelemetry("DEVTOOLS_WEBIDE_SIMULATOR_CONNECTION_RESULT", [0,1,0]);
        checkTelemetry("DEVTOOLS_WEBIDE_TIME_ACTIVE_SECONDS", null, "hasentries");
        checkTelemetry("DEVTOOLS_WEBIDE_USB_CONNECTION_RESULT", [0,1,0]);
        checkTelemetry("DEVTOOLS_WEBIDE_WIFI_CONNECTION_RESULT", [0,1,0]);
      }

      window.onload = function() {
        SimpleTest.waitForExplicitFinish();

        let win;

        SimpleTest.registerCleanupFunction(() => {
          Task.spawn(function*() {
            if (win) {
              yield closeWebIDE(win);
            }
            DebuggerServer.destroy();
            yield removeAllProjects();
          });
        });

        Task.spawn(function*() {
          Cu.import("resource://gre/modules/devtools/dbg-server.jsm");

          if (!DebuggerServer.initialized) {
            DebuggerServer.init();
            DebuggerServer.addBrowserActors();
          }

          startTelemetry();

          // Cycle once, so we can test for multiple opens
          yield cycleWebIDE();

          win = yield openWebIDE();
          // Wait a bit, so we're open for a non-zero time
          yield waitForTime(TOOL_DELAY);
          addFakeRuntimes(win);
          yield addTestApp(win);

          // Each one should log a connection result and non-zero connection
          // time
          yield connectToRuntime(win, "usb");
          yield waitForTime(TOOL_DELAY);
          yield connectToRuntime(win, "wifi");
          yield waitForTime(TOOL_DELAY);
          yield connectToRuntime(win, "simulator");
          yield waitForTime(TOOL_DELAY);
          yield connectToRuntime(win, "other", 0 /* remote */);
          yield waitForTime(TOOL_DELAY);
          yield connectToRuntime(win, "other", 1 /* local */);
          yield waitForTime(TOOL_DELAY);
          yield connectToRuntime(win, "other", 2 /* other */);
          yield waitForTime(TOOL_DELAY);
          yield closeWebIDE(win);

          checkResults();

          SimpleTest.finish();
        });
      }
    </script>
  </body>
</html>
